Refactor and reorganize minishell codebase
- Removed obsolete files: exec/read.c and exec/read_process.c.
- Introduced new files for handling command execution, pipelines, and redirections.
- Implemented exec_and_or functionality to manage logical AND/OR operations in command execution.
- Enhanced built-in command execution with exec_builtins.c.
- Added comprehensive pipeline handling in exec_pipe.c.
- Developed redirection handling in exec_redirection.c.
- Created heredoc management in herdoc_handler.c and herdoc_utils.c.
- Updated header file minishell.h to reflect new function declarations and structures.
- Improved input processing and token handling in read_process.c.
- Ensured proper memory management and error handling throughout the codebase.
