+++ builtins/export/print_env.c
@@ -3,14 +3,14 @@
 /*                                                        :::      ::::::::   */
 /*   print_env.c                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: malrifai <malrifai@student.42.fr>          +#+  +:+       +#+        */
+/*   By: eaqrabaw <eaqrabaw@student.42amman.com>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/02/22 18:00:15 by malrifai          #+#    #+#             */
-/*   Updated: 2025/04/30 16:31:45 by malrifai         ###   ########.fr       */
+/*   Updated: 2025/05/04 07:48:48 by eaqrabaw         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "../../includes/minishell.h"
+#include "minishell.h"
 
 void	is_larger(int *swapped, t_env **ptr)
 {
